FixChain/requirements.txt
jinja2
uvicorn
fastapi
pydantic
requests
python-dotenv
pymongo>=4.6,<5
python-multipart
google-genai
mcp>=1.0.0
aiofiles
aiohttp